bugfix : fix advanced mode and camera recentering
This commit addresses the bug of the camera recentering and the advanced being not persistent across screens navigations. This commit also proceeds to refactor the logic to MapViewModel and correct a minor issue.
